Time: 0 ms (100.00%) | Memory: 7.2 MB (37.27%) - LeetSync

+++ 191-number-of-1-bits/number-of-1-bits.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int hammingWeight(uint32_t n) {
-        return __builtin_popcount(n);
+        return __builtin_popcount((int)n);
     }
 };
